libstdc++: Remove redundant cast in floating_from_chars.cc
In r15-7647-g32457bc25fea80 I added a cast and also changed the type of
the variable, making the cast redundant. This removes the cast.

libstdc++-v3/ChangeLog:

	* src/c++17/floating_from_chars.cc (__floating_from_chars_hex):
	Remove redundant cast.

+++ libstdc++-v3/src/c++17/floating_from_chars.cc
@@ -1103,8 +1103,7 @@ namespace
 	// If the leading hexit is not '1', shift MANTISSA to make it so.
 	// This normalizes input like "4.08p0" into "1.02p2".
 	const unsigned leading_hexit = mantissa >> mantissa_bits;
-	const int leading_hexit_width
-	  = __bit_width((unsigned)leading_hexit); // FIXME: optimize?
+	const int leading_hexit_width = __bit_width(leading_hexit); // FIXME: optimize?
 	__glibcxx_assert(leading_hexit_width >= 1 && leading_hexit_width <= 4);
 	shift_mantissa(leading_hexit_width - 1);
 	// After this adjustment, we can assume the leading hexit is '1'.
